requirements.txt
discord.py-self==2.0.0
spotipy==2.24.0